requirements.txt
google-generativeai
pypdf
Pillow
sentence-transformers
scikit-learn
streamlit
